pyproject.toml
[tool.poetry]
name = "todo-api"
version = "0.1.0"
description = ""
authors = ["kirillketrik <kirillketrik34@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
tortoise-orm = "^0.25.1"
uvicorn = "^0.35.0"
pydantic-settings = "^2.10.1"
passlib = "^1.7.4"
python-jose = "^3.5.0"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
